Update installLogging.ts
Was there a reason why there's a ternary again for isDev? Due to it being inside an else statement for !isDev, isDev will never be true in the ternary.
The choice was either get rid of the if/else statement or delete the ternary. I assume some people don't want even tiny logging in dev; hence, I kept the if/else but removed the ternary.